[FEATURE] introduce `compact-mode`
To tailor to very minimalistic needs regarding design, this introduces a
compact mode that strips out most of the design elements and makes
maximum use of the screen space. This is also intended to work well on
small devices or using a large browser zoom.

The mode can be enabled by browsing `compact.html` instead of
`index.html`

Requested-By: Fabian Schneider <faebl.taylor@pm.me>