!fixup
Decouples the compact mode from the standard HTML site. This is done to
prevent changes in the normal mode to meddle with the compact mode and
to force changes in the normal mode to be thought through und
implemented separately in the compact mode to keep features/code
consistent.
Same for compact.css

Note: this change is not complete/fully functional yet.